Add local user if not exists
The a local WordPress user does not exist while logging in it will not be created.
The WordPress setting to allow registration must be checked in Settings > General
in order for the new user to be registered. Otherwise the user attempting to login
without a local account will receive a message notifying them that the account
does not exist.