Isolates Firebase\JWT class with namespace change
 * Changes 'Firebase\JWT' namespace to 'AADSSO\Firebase\JWT'.
 * Closes issue #88, and is a long-term solution for #28.

Note: While I'm not particularly keen on changing the code for a
third-party library (even if it's just a namespace change), this method
both prevents class name (and namespace) collision, and ensures the
expected version of the library is used. The latter point is
particularly important for security-sensitive classes such as
Firebase/JWT, which does JWT token validation.