Fix issue when for unprovisioned user not in group
When a user does not exist in WordPress, auto-provisioning is enabled, group-to-role mapping is enabled, and the signed-in user does not belong to any of the defined groups, we should deny access if there is no default user role defined. If there is a default user role defined, we allow access. This commit addresses issue where un-provisioned users would always get access denied, even if there was a default role defined.